forpython/requirements.txt
cachetools==4.2.1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
croniter==1.0.2
cycler==0.10.0
dnspython==2.1.0
google-api-core==1.26.3
google-api-python-client==2.1.0
google-auth==1.28.0
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.4.4
googleapis-common-protos==1.53.0
httplib2==0.19.1
idna==2.10
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
natsort==7.1.0
numpy==1.19.5
oauthlib==3.1.0
packaging==20.9
pandas==1.2.1
Pillow==8.1.0
protobuf==3.15.7
pyasn1==0.4.8
pyasn1-modules==0.2.8
pymongo==3.11.2
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.5
requests==2.25.1
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.15.0
tqdm==4.56.0
transliterate==1.10.2
uritemplate==3.0.1
urllib3==1.26.2
python-dotenv==0.19.1
